Split files and attempt to write tests
