+++ frontend/src/metabase/visualizations/components/ObjectDetail/ObjectDetail.styled.tsx
@@ -2,6 +2,7 @@ import styled from "@emotion/styled";
 import Modal from "metabase/components/Modal";
 import { breakpointMinMedium } from "metabase/styled-components/theme/media-queries";
 import { color } from "metabase/lib/colors";
+import Button from "metabase/core/components/Button";
 import TableFooter from "../TableSimple/TableFooter";
 
 interface ObjectDetailContainerProps {
@@ -50,15 +51,22 @@ export const ObjectRelationships = styled.div`
 `;
 
 export const CloseButton = styled.div`
-  display: flex;
+  //display: flex;
   margin-left: 1rem;
   padding-left: 1rem;
   border-left: 1px solid ${color("border")};
 
+  // ${breakpointMinMedium} {
+  //position: absolute !important;
+  //top: -9999px !important;
+  //left: -9999px !important;
+  display: none;
+  //}
+`;
+
+export const CloseButtonTest = styled(Button)`
   ${breakpointMinMedium} {
-    position: absolute !important;
-    top: -9999px !important;
-    left: -9999px !important;
+    display: none;
   }
 `;
